Added logout feature to users
In logout route we are accessing the token which is passed in the request and running a filter function which will return true for all other tokens, except of the one which matches with out request token. Hence removing that token from token array in user model.
Additionally added a logoutAll route in which even if user is logged in across multiple devices and has many jwt's he'll be logged out of the all and token array will become empty